+++ sdk/paloaltonetworks/azure-mgmt-paloaltonetworksngfw/azure/mgmt/paloaltonetworksngfw/aio/operations/_operations.py
@@ -11481,7 +11481,7 @@ class PaloAltoNetworksCloudngfwOperationsOperations:  # pylint: disable=name-too
 
     @distributed_trace_async
     async def create_product_serial_number(self, **kwargs: Any) -> _models.ProductSerialNumberRequestStatus:
-        """create_product_serial_number.
+        """A long-running provider action.
 
         :return: ProductSerialNumberRequestStatus. The ProductSerialNumberRequestStatus is compatible
          with MutableMapping
@@ -11545,7 +11545,7 @@ class PaloAltoNetworksCloudngfwOperationsOperations:  # pylint: disable=name-too
 
     @distributed_trace_async
     async def list_cloud_manager_tenants(self, **kwargs: Any) -> _models.CloudManagerTenantList:
-        """list_cloud_manager_tenants.
+        """A long-running provider action.
 
         :return: CloudManagerTenantList. The CloudManagerTenantList is compatible with MutableMapping
         :rtype: ~azure.mgmt.paloaltonetworksngfw.models.CloudManagerTenantList
@@ -11608,7 +11608,7 @@ class PaloAltoNetworksCloudngfwOperationsOperations:  # pylint: disable=name-too
 
     @distributed_trace_async
     async def list_product_serial_number_status(self, **kwargs: Any) -> Optional[_models.ProductSerialNumberStatus]:
-        """list_product_serial_number_status.
+        """A long-running provider action.
 
         :return: ProductSerialNumberStatus or None. The ProductSerialNumberStatus is compatible with
          MutableMapping
@@ -11674,7 +11674,7 @@ class PaloAltoNetworksCloudngfwOperationsOperations:  # pylint: disable=name-too
 
     @distributed_trace_async
     async def list_support_info(self, **kwargs: Any) -> _models.SupportInfoModel:
-        """list_support_info.
+        """A long-running provider action.
 
         :return: SupportInfoModel. The SupportInfoModel is compatible with MutableMapping
         :rtype: ~azure.mgmt.paloaltonetworksngfw.models.SupportInfoModel